Pull search and generic views into project app
