+++ ncdr_reference/urls.py
@@ -15,9 +15,23 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import include, path
+from django.views.decorators.cache import never_cache
+
+from .views import AddMany, Delete, Edit, List, Search, SearchRedirect, Unpublished
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+
     path('', include("csv_schema.urls")),
+
+    path('form/<slug:model_name>/unpublished/', never_cache(Unpublished.as_view()), name="unpublished_list"),
+    path('form/<slug:model_name>/add/', AddMany.as_view(), name="add_many"),
+    path('form/<slug:model_name>/edit/<int:pk>/', Edit.as_view(), name="edit"),
+    path('form/<slug:model_name>/delete/<int:pk>/', Delete.as_view(), name="delete"),
+    path('form/<slug:model_name>/', never_cache(List.as_view()), name="edit_list"),
+
+    path('search/', SearchRedirect.as_view(), name="search_redirect"),
+    path('search/<slug:model_name>/', Search.as_view(), name="search"),
+
     path('accounts/', include('django.contrib.auth.urls')),
 ]
